refactor: split by files
(cherry picked from commit e1fe19bc2a033f3dc60150ea1f9547d38567ba34)
