+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "node ./src/index.js",
+    "start": "node src/commandHandler.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
